+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Core
 ~~~~
 * Multiply photon flux with detector time step in photon generation models.
   (See `!305 <https://gitlab.com/esa/pyxel/-/merge_requests/305>`_).
+* Initialize Photon class in detector reset function instead in models.
+  (See `!309 <https://gitlab.com/esa/pyxel/-/merge_requests/309>`_).
 
 Documentation
 ~~~~~~~~~~~~~